debian-autoconfiguredesktop
I use this script to install some of the stuff I like on my desktops and laptops running anything Debian-based. I decided to put it up here because it's nice and easy for me to come grab it here and use it on any computer I have. I also have cleverly timed "sleep" commands so I can have my tea steep, coming back to a prompt with my name and a "$" waiting for me, with a nice cup of tea.